Added some lazy method bodies to the player and grid class, Display now draws pieces in the bank and grid.
I'm assuming that what I did is only going to be temporary. When we can
actually talk about how we're going to do it, then we will do that.

There is a slight issue- that is being that when pieces are added and
removed from the grid, the arrays get scrambled, so when we draw it with
the pieces they will not really be in order.
That being said I have a solution: have the PuzzlePieceImage.java class
extend the PuzzlePiece.java class.

Please tell me if I am wrong, dumb, or both. 
Thanks.