Update Display.java
Made the frame nonresizeable can change that real quick if we don't want it.
Changed it so that I could throw a box around the grid area because I couldn't figure out where the grid was height wise by reading code. If you want to make it resizeable, go for it.